requirements.txt
asgiref==3.4.1
Django==4.0
django-widget-tweaks==1.4.9
Pillow==8.4.0
python-dotenv==0.19.2
sqlparse==0.4.2
